refactor(docs): reorganize markdown utilities and update imports for clarity
